Address Linard's review comments from Feb 2 on PR #983 (#1067)
- GobraParser.g4: align the order of the spec modifier alternatives with
  the order of the attributes in the rule's returns clause
- Introduce a TerminationTyping trait collecting the termination-measure
  helpers (moved from GhostMiscTyping) so that all well-definedness
  checks reason about termination consistently
- Strengthen the atomic-member termination check: conditional measures
  alone no longer count as a termination guarantee; a non-conditional
  measure is required
- critical statements: accept expressions whose underlying type is
  pred() instead of requiring exactly pred()
- critical statements: classify body statements into ghost stmts,
  non-ghost atomic ops, and invalid stmts, and report errors at the
  offending statement with clearer messages
