+++ src/main/scala/viper/gobra/frontend/info/implementation/typing/ghost/GhostMemberTyping.scala
@@ -114,11 +114,10 @@ trait GhostMemberTyping extends BaseTyping { this: TypeInfoImpl =>
       "Ghost members cannot be marked as atomic.",
       spec.isAtomic && isGhost
     )
-    val decreases = spec.terminationMeasures.nonEmpty
     val atomicMethodsTerminate = error(
       member,
-      "Atomic members must be specified as terminating with a decreases-clause.",
-      spec.isAtomic && !decreases
+      "Atomic members must be guaranteed to terminate on every call. Their specification must contain a non-conditional decreases-clause.",
+      spec.isAtomic && !measuresGuaranteeTermination(spec.terminationMeasures)
     )
     atomicsAreAbstract ++ atomicsAreActual ++ atomicMethodsTerminate
   }
